requirements.txt
fastapi
uvicorn
pdfplumber
PyMuPDF
pytesseract
Pillow
transformers
torch
spacy
scikit-learn
pandas
numpy
python-dotenv
